+++ Draw.js
@@ -2,11 +2,14 @@
 
 /* globals
 canvas,
+foundry,
 PIXI,
 CONFIG
 */
 
 "use strict";
+
+import { GEOMETRY_CONFIG } from "./const.js";
 
 // Draw class for drawing shapes; primarily for debugging
 
@@ -204,3 +207,6 @@ export class Draw {
     this.g.clear();
   }
 }
+
+GEOMETRY_CONFIG.Draw ??= Draw;
+