+++ PIXI/Rectangle.js
@@ -3,7 +3,7 @@ PIXI
 */
 "use strict";
 
-import { Matrix } from "../Matrix.js";
+import "../Matrix.js";
 import { cutawayBasicShape, cutawayBasicIntersections } from "../util.js";
 
 export const PATCHES = {};
@@ -433,8 +433,8 @@ function rotateAroundCenter(rotation = 0) {
   }
 
   // For all other rotations, translate center to 0,0, rotate, and then invert the translation.
-  const tMat = Matrix.translation(-center.x, -center.y);
-  const rMat = Matrix.rotationZ(Math.toRadians(rotation));
+  const tMat = CONFIG.GeometryLib.Matrix.translation(-center.x, -center.y);
+  const rMat = CONFIG.GeometryLib.Matrix.rotationZ(Math.toRadians(rotation));
   const M = tMat.multiply3x3(rMat).multiply3x3(tMat.invert);
   const pts = [...this.iteratePoints({ close: true })];
   const tPts = pts.map(pt => M.multiplyPoint2d(pt));